+++ extensions/calendly/webhooks/index.ts
@@ -1,6 +1,8 @@
 import { eventCreated } from './eventCreated'
-import { eventCanceled } from './eventCancelled'
+import { eventCanceled } from './eventCanceled'
+import { eventRescheduled } from './eventRescheduled'
 export type { EventCreated } from './eventCreated'
-export type { EventCanceled } from './eventCancelled'
+export type { EventCanceled } from './eventCanceled'
+export type { EventRescheduled } from './eventRescheduled'
 
-export const webhooks = [eventCreated, eventCanceled]
+export const webhooks = [eventCreated, eventCanceled, eventRescheduled]