+++ extensions/calendly/CHANGELOG.md
@@ -2,4 +2,4 @@
 
 ## November 3, 2023
 
-- Calendly webhooks for event created and event canceled are now supported.
+- Calendly webhooks for event created, canceled and rescheduled are now supported.